Add files via upload
This article covers in depth analysis to optimize use of InMemory in Oracle. 
With the latest Oracle releases it has given a significant performance boost with many features like InMemory. InMemory is basically an area in Memory (RAM) that is reserved for use to PERSISTENTLY store data from tables. The data in memory is pushed to InMemory manually and it remains there till we pull it from InMemory giving the developer/dba/Performance tuners a unique opportunity to select the data that they feel needs to be retrieved fast (rather than relying on Oracle engine to select data based on hotness like in case of buffer)
When we run a query the optimizer than looks for the data in InMemory, if its able to find the data in InMemory, it pulls the data from InMemory before looking for the data in disks. The IO operations hence are approximately ten times faster and data is retrieved in seconds as compared to minutes needed from disk.
However there are limitations and the limitation here is the size of inmemory that is there. If you have unlimited inmemory available then we can simply put all the tables into inmemory and the task of optimization is over.

But in real world scenario the Performance tuning team has a task at hand to optimally use the InMemory available and bring out the maximum benefit from the least amount of inmemory. 

Lets take an example, if a table has 100 columns and size of table is 7GB (after compression) and only 10 columns are being used in the reports, putting the whole table into InMemory would not make sense, it means we are using complete 7GB of InMemory. However we are actually using just 10 columns and rest 90 columns which are sitting in inmemory are actually wasting the space. The optimal way to use the InMemory would be to just put these 10 columns to inmemory and pull out rest 90 columns. The size used in memory now would be 0.7GB only. This way we are still getting the same benefit we were getting earlier but at a fraction of memory being used. 

Now that we have saved 90% of the space in InMemory, lets say there is another query that gets executed using 11th column, now this will not be served from InMemory only and the query wont perform well. Hence its a tuners job to constantly monitor the SQLs being directed to InMemory tables and constantly evolve the best set of columns that has to be put InMemory maximizing the benefits.

Automation:
Imagine doing this for all the queries that the frontend is issuing day in day out on hundreds of tables. This definitely needs some utility like the AWR etc. but there isnt one out yet from Oracle.
For this purpose i have designed a query that scans through the metadata from obiee queries that got executed in the past (using the usage tracking tables), scans through the tables being used, filtering out the columns being used in those queries. It then combines these findings to present a cumulative list of columns that need to be put inmemory that are being used in the reports. The utility further scans the InMemory area and fetches whats currently inmemory and also creates "alter table..." statements that can directly be used to put the corresponding columns into inmemory. All this can be done in minutes, constantly on a daily basis , ever evolving for all tables, for all columns.